ready: Move from module config to env vars
Using the new /ready listener/endpoint for the healcheck means the go code
will be involved at some point.
Right now, the config for the ready endpoint is based on 3 fields:
ready_port
ready_ip
ready_log

In the current code, those 3 fields are part of the ambassador module config.

When the go code is involved, there will be a race condition on "when" those values
are applied/taken into account between the go process and the python+envoy process.
So that might be a problem.
What could happen is the following:
The ambassador module is changed and ready_port is changed for instance.
t0: the go process gets the change and starts to process it
t0: the python process gets the change and starts to process it.
t10: the go process has processed the change and starts to ping the new port for /ready endpoint.
t100: the python process and envoy has changed the port.

During t10 to t100, the go process /ready check will fail, because the refresh of the config
between the go process and the python+envoy process are not synced.
That might be a problem and generate problems during updates, with the pod becoming not ready
and k8s not sending traffic to it anymore.

Alternative option:
Move the 3 fields out of the ambassador module, define some defaults, that can only be changed with env vars.
The 3 env vars are:
AMBASSADOR_READY_PORT
AMBASSADOR_READY_IP
AMBASSADOR_READY_LOG

This way, the config cannot change between updates, it's setup at startup.

Signed-off-by: Fabrice Rabaute <fabrice@arista.com>